+++ cli/cmd/read.go
@@ -51,6 +51,12 @@ The command accepts other arguments too.`,
 		watch, _ := cmd.Flags().GetBool("watch");
 		noLog, _ := cmd.Flags().GetBool("no-log");
 		interval, _ := cmd.Flags().GetInt("interval")
+		unit, _ := cmd.Flags().GetString("unit")
+
+		if (unit != "c" && unit != "f" && unit != "k") {
+			print(color.MagentaString("The --unit (-u) flag only accepts: 'c', 'f', 'k',\nbut you entered \"%s\".\r\n\r\nReminder that the default is 'c'.\r\n", unit))
+			return
+		}
 
 		// exit gracefully on ctrl+c
 		var sigChan = make(chan os.Signal, 1)
@@ -62,7 +68,7 @@ The command accepts other arguments too.`,
 		}
 		port, err := serial.Open(serialPort, mode)
 		if err != nil {
-			var errorString string = color.HiRedString("\r\nThere was an error while trying to connect to the ODSCI probe.\r\nThe serial port you entered may be incorrect.\r\nTo scan for serial ports on your computer, run ") + color.HiMagentaString("'odsci scan'") + color.HiRedString(".\r\n\r\nError details:\r\n\r\n")
+			var errorString string = color.HiRedString("\r\nThere was an error while trying to connect to the ODSCI probe.\r\nThe serial port you entered may be incorrect." + color.HiRedString("\r\n\r\nError details:\r\n\r\n"))
 			print(errorString)
 			log.Fatal(err)
 		}
@@ -86,20 +92,40 @@ The command accepts other arguments too.`,
 			if (noLog) {
 				fmt.Print(color.HiYellowString("You are using the \"--no-log\" flag. If the CLI looks like it has frozen, it hasn't.\r\nThe temperature is just not updating.\r\n\r\n"))
 			}
+			var prevTemperature float64;
+			var firstLog bool = true;
 			for (true) {
+				_, raw_temp := utils.ReadTemperature(port, scanner)
+				var temp_to_print string
+				var difference = raw_temp - prevTemperature;
+				if (!firstLog) {
+					switch unit {
+						case "c":
+							temp_to_print = fmt.Sprintf("%+0.2fºC (%+0.2fºC)", raw_temp, difference)
+						case "f":
+							temp_to_print = fmt.Sprintf("%+0.2fºF (%+0.2fºF)", utils.ConvertCelsiusToFahrenheit(raw_temp), utils.ConvertCelsiusToFahrenheit(difference))
+						case "k":
+							temp_to_print = fmt.Sprintf("%+0.2fºK (%+0.2fºK)", utils.ConvertCelsiusToKelvin(raw_temp), utils.ConvertCelsiusToFahrenheit(difference))
+					}
+				} else {
+					switch unit {
+						case "c":
+							temp_to_print = fmt.Sprintf("%+0.2fºC", raw_temp)
+						case "f":
+							temp_to_print = fmt.Sprintf("%+0.2fºF", utils.ConvertCelsiusToFahrenheit(raw_temp))
+						case "k":
+							temp_to_print = fmt.Sprintf("%+0.2fºK", utils.ConvertCelsiusToKelvin(raw_temp))
+					}
+				}
 				if (!noLog) {
 					timestamp := time.Now().UTC().Format("15:04:05")
-					fmt.Printf("[%s]: %s\r\n",timestamp, func() string {
-						s, _ := utils.ReadTemperature(port, scanner)
-						return s
-					}())
+					fmt.Printf("[%s]: %s\r\n",timestamp, temp_to_print)
 				} else {
 					timestamp := time.Now().UTC().Format("15:04:05")
-					fmt.Printf("\r[%s]: %-10s",timestamp, func() string {
-						s, _ := utils.ReadTemperature(port, scanner)
-						return s
-					}())
+					fmt.Printf("\r[%s]: %-10s",timestamp, temp_to_print)
 				}
+				prevTemperature = raw_temp;
+				if (firstLog == true) { firstLog = false }
 				time.Sleep(time.Duration(interval) * time.Second)
 			}
 		} else {
@@ -117,4 +143,5 @@ func init() {
 	readCmd.Flags().BoolP("watch", "w", false, `Continuously watch the temperature reading`)
 	readCmd.Flags().IntP("interval", "i", 1, "Interval for watching, if using the --watch flag")
 	readCmd.Flags().Bool("no-log", false, "No log: applies only if --watch is being used and does't log previous values")
+	readCmd.Flags().StringP("unit", "u", "c", "Select a unit: 'c' for Celsius, 'f' for Fahrenheit or 'k' for Kelvin")
 }