+++ cli/utils/fahrenheit_kelvin.go
@@ -1,0 +1,9 @@
+package utils
+
+func ConvertCelsiusToFahrenheit(celsius float64) float64 {
+	return celsius * 1.8 + 32
+}
+
+func ConvertCelsiusToKelvin(celsius float64) float64 {
+	return celsius + 273.15
+}